TrustXCM.java: set name, state and some empty strings for the webui
The attribute names was not set for the container object. Further, the WebUI
seems not to be happy with null attributes in a container object. Thus, this
patch sets the relevant String attributes to "" instead of leaving them
uninitialized (null). To provide correct state changes in webui a simple
mapper to the "Up"/"Exited" String representation of a Container state is
introduced.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>