Report renewal errors from the EST to the user
This makes use of the default spring behaviour for unhandled exceptions.
Note that to report a useful message to the user, the application setting
`server.error.include-message` must be set to true.
If not, the UI will just show a generic message with the suggestion to
look into the connector log.